Move request-specific code into its own controller
