+++ dashboard/app/models/ability.rb
@@ -490,9 +490,14 @@ class Ability
         user.has_ai_tutor_access?
       end
 
-      can [:log_chat_event, :start_chat_completion, :chat_request, :find_toxicity], :aichat do
+      can [:start_chat_completion, :chat_request], :aichat_request do
         user.teacher_can_access_ai_chat? || user.student_can_access_ai_chat?
       end
+
+      can [:log_chat_event, :find_toxicity], :aichat do
+        user.teacher_can_access_ai_chat? || user.student_can_access_ai_chat?
+      end
+
       # Additional logic that confirms that a given teacher should have access
       # to a given student's chat history is in aichat_controller.
       can :student_chat_history, :aichat do
