+++ dashboard/config/routes.rb
@@ -1200,10 +1200,11 @@ Dashboard::Application.routes.draw do
 
     post '/openai/chat_completion', to: 'openai_chat#chat_completion'
 
+    post '/aichat/start_chat_completion', to: 'aichat_requests#start_chat_completion'
+    get '/aichat/chat_request/:id', to: 'aichat_requests#chat_request'
+
     post '/aichat/log_chat_event', to: 'aichat#log_chat_event'
     get '/aichat/student_chat_history', to: 'aichat#student_chat_history'
-    post '/aichat/start_chat_completion', to: 'aichat#start_chat_completion'
-    get '/aichat/chat_request/:id', to: 'aichat#chat_request'
     get '/aichat/user_has_access', to: 'aichat#user_has_access'
     post '/aichat/find_toxicity', to: 'aichat#find_toxicity'
